+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 `1.0.44`
 
+- Added `Throwables.find(Throwable, Class)` to return the first throwable in a cause chain matching a given type.
+
 ---
 
 `1.0.43`